+++ src/bucket.h
@@ -2,14 +2,13 @@
 #define __BUCKET__
 
 #include <stddef.h>
-#include <stdint.h>
 
 typedef struct {
-    uint32_t* elems;
+    double* elems;
     size_t n_elems;
     size_t capacity;
 } Bucket;
-typedef size_t (*BucketIdx)(uint32_t);
+typedef size_t (*BucketIdx)(double);
 
 typedef enum {
     PUSH_SUCCESSFUL,
@@ -21,7 +20,7 @@ typedef enum {
     FAILURE,
 } BucketSortStatus;
 
-BucketStatus bucket_push(Bucket* bucket, uint32_t elem);
+BucketStatus bucket_push(Bucket* bucket, double elem);
 
 Bucket bucket_init(size_t capacity);
 Bucket* create_buckets(size_t n, size_t capacity);
@@ -29,6 +28,6 @@ Bucket* create_buckets(size_t n, size_t capacity);
 void bucket_free(Bucket* bucket);
 void destroy_buckets(Bucket* buckets, size_t n);
 
-BucketSortStatus bucket_sort(uint32_t* array, size_t n_elems, Bucket* buckets, size_t n_buckets, BucketIdx bucket_idx);
+BucketSortStatus bucket_sort(double* array, size_t n_elems, Bucket* buckets, size_t n_buckets, BucketIdx bucket_idx);
 
 #endif
